+++ Framework11/vprogeditor/node.cpp
@@ -267,7 +267,7 @@ bool cNode::RenderStringPin(cEditManager &editMgr
 {
 	const string scopeName =
 		common::script::cSymbolTable::MakeScopeName(
-			m_name.c_str(), m_id.Get());
+			m_name.c_str(), (int)m_id.Get());
 
 	common::script::sVariable *varInfo =
 		editMgr.m_symbTable.FindVarInfo(scopeName, pin.name.c_str());
@@ -310,7 +310,7 @@ bool cNode::RenderEnumPin(cEditManager &editMgr
 {
 	const string scopeName =
 		common::script::cSymbolTable::MakeScopeName(
-			m_name.c_str(), m_id.Get());
+			m_name.c_str(), (int)m_id.Get());
 	common::script::sVariable *varInfo =
 		editMgr.m_symbTable.FindVarInfo(scopeName, pin.name.c_str());
 	if (!varInfo)
